Added Email Sender Code and Tested APIs

+++ Futurix-Bank-Backend/src/main/java/com/futurix/repositories/AccountRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.futurix.entities.TblAccount;
 
-public interface AccountRepo extends JpaRepository<TblAccount, Long> {
+public interface AccountRepo extends JpaRepository<TblAccount, Integer> {
 
 }
